clase6/pyproject.toml
[tool.poetry]
name = "clase6"
version = "0.1.0"
description = ""
authors = ["RArielAlonso <rariel.alonso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.5,<3.13"
pandas = "^2.1.1"
pyarrow = "^13.0.0"
fastparquet = "^2023.8.0"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.11.0"
gcsfs = "^2023.9.2"
google-cloud-bigquery = "^3.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
